requirements.txt
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
Flask==2.2.2
idna==3.4
importlib-metadata==5.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
requests==2.28.1
urllib3==1.26.13
Werkzeug==2.2.2
zipp==3.11.0

python-dateutil~=2.8.2
pip~=22.3
wheel~=0.37.0
setuptools~=58.0.4
six~=1.15.0
future~=0.18.2